several code tweaks and improvements for core framework

+++ inc/shapers/circle.h
@@ -30,7 +30,7 @@ public:
     };;
 
     // optional
-    void onClick(int x, int y) override;
+    void onClick(float x, float y) override;
 
     // required
     void _activate() override {};
