+++ inc/shapers/star.h
@@ -19,7 +19,7 @@ class StarTool : public aiero::Tool {
         };
 
         // optional
-        void onClick(int x, int y) override;
+        void onClick(float x, float y) override;
 
         // required
         void _activate() override {};
