+++ inc/tools/paintbrush.h
@@ -19,9 +19,9 @@ class PaintBrushTool : public aiero::Tool {
         }; // to be implemented in the source files
 
         // optional
-        void onClick(int x, int y) override;
-        // void onMouseDown(int x, int y) override;
-        // void onMouseUp(int x, int y) override;
+        void onClick(float x, float y) override;
+        // void onMouseDown(float x, float y) override;
+        // void onMouseUp(float x, float y) override;
 
         // required
         void _activate() override {
